fix(preferences): isolate temp context mode constants
