+++ tests/entrypoints/background/tempWindowPoolNativeCheckin.test.ts
@@ -2,7 +2,6 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
 
 import { RuntimeActionIds } from "~/constants/runtimeActions"
 import { SITE_TYPES } from "~/constants/siteType"
-import { TEMP_CONTEXT_MODES } from "~/services/preferences/userPreferences"
 import { AuthTypeEnum } from "~/types"
 
 const originalBrowser = (globalThis as any).browser
@@ -117,14 +116,13 @@ describe("tempWindowPool native check-in page action", () => {
     vi.doMock("~/services/preferences/userPreferences", () => ({
       DEFAULT_PREFERENCES: {
         tempWindowFallback: {
-          tempContextMode: TEMP_CONTEXT_MODES.Tab,
+          tempContextMode: "tab",
         },
       },
-      TEMP_CONTEXT_MODES,
       userPreferences: {
         getPreferences: vi.fn().mockResolvedValue({
           tempWindowFallback: {
-            tempContextMode: TEMP_CONTEXT_MODES.Tab,
+            tempContextMode: "tab",
           },
         }),
       },
