+++ tests/entrypoints/background/tempWindowPoolSuspendCleanup.test.ts
@@ -1,7 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest"
 
 import { RuntimeActionIds } from "~/constants/runtimeActions"
-import { TEMP_CONTEXT_MODES } from "~/services/preferences/userPreferences"
 
 const originalBrowser = (globalThis as any).browser
 
@@ -63,14 +62,13 @@ describe("cleanupTempContextsOnSuspend", () => {
     vi.doMock("~/services/preferences/userPreferences", () => ({
       DEFAULT_PREFERENCES: {
         tempWindowFallback: {
-          tempContextMode: TEMP_CONTEXT_MODES.Tab,
+          tempContextMode: "tab",
         },
       },
-      TEMP_CONTEXT_MODES,
       userPreferences: {
         getPreferences: vi.fn().mockResolvedValue({
           tempWindowFallback: {
-            tempContextMode: TEMP_CONTEXT_MODES.Tab,
+            tempContextMode: "tab",
           },
         }),
       },
